Combine tone detection and generation into single tone_detect.py
- Merge tone_generation.py and detection.py into tone_detect.py
- Both functions are closely related and share dependencies
- Reduces module count and simplifies imports
- Update imports in __init__.py and audio_setup.py
